The filtering of mediator conflicts will be changed
1) Изменена фильтрация конфликтов медиатора;
2) В сайдбаре и меню пользователя выполнена
   подсветка кнопок соответствующих страниц;
3) Добавлен вывод текста об отсутствующих
   конфликтах на страницах manage-jobs
   медиатора и юзера;
4) Добавлена пагинация на этих страницах;

+++ mediators_website/conflict/models.py
@@ -85,7 +85,7 @@ class Conflict(models.Model):
     # prise = models.TextField(choices=PriseChoices.choices,
     #                          default=PriseChoices.CHOOSE,
     #                          verbose_name=_("Цена"))
-    fixed_price = models.FloatField(verbose_name=_("Цена"))
+    fixed_price = models.PositiveIntegerField(verbose_name=_("Цена"))
     decide_time = models.TextField(choices=DecidedTime.choices,
                                    default=DecidedTime.CHOOSE,
                                    verbose_name=_("Время на решение"))
